The following use cases may appear after programming the firmware: 1. The filesystem is not programmed, because the user forgot it. 2. The filesystem itself is valid, but not its content, because the user forgot to program it. 3. The filesystem is valid, but its content is from a older Pixelix version, because the user forgot to update it.
Solutions:
to 1.) If filesystem can not be mounted, a low level error code is displayed.
to 2.) If filesystem is mounted, but the version.json is missing, a low level error code is displayed.
to 3.) If filesystem is mounted, version.json available, but the version doesn't match, a warning is displayed.

---

Additional the version bumped to v4.1.2