Some configuration parameters can now be selected in a separate GUI. The possible parameters are in the configuration database in JSON format. The scripts/configure.py is called as pre-type script in every pio environment and will append/replace build parameters.
The selected configuration parameters will be stored in ./config.json in the root directory. Once its there, the GUI won't appear anymore. To select a different setup, remove the ./config.json file and start the build process again.

In principle it is possible to change the ./config.json by hand, BUT keep in mind to clear all previously built objects! pio can not detect these change and won't start a clean build run.