WEB: Optimize GETDISP command
Add a simple "compression" to framebuffer uploads. The unused
upper 8 bits of an RGB value are now a repeat count. It tells how
often the same color shall be used on the pixels that follow.

Example:
To upload an empty black 32x8 frame buffer, only a
single 0xFF000000 data value will be send.